+++ src/Greenleaf/Action/ByMethodTrait.php
@@ -10,7 +10,9 @@ declare(strict_types=1);
 namespace DecodeLabs\Greenleaf\Action;
 
 use DecodeLabs\Greenleaf\ActionTrait;
+use DecodeLabs\Greenleaf\Route;
 use DecodeLabs\Harvest;
+use DecodeLabs\Harvest\Request as HarvestRequest;
 use DecodeLabs\Singularity\Url\Leaf as LeafUrl;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
@@ -31,11 +33,7 @@ trait ByMethodTrait
         $method = $request->getMethod();
 
         if (!method_exists($this, $method)) {
-            if ($method === 'HEAD') {
-                $method = 'GET';
-            } else {
-                return $this->handleUnknownMethod($method);
-            }
+            return $this->handleUnknownMethod($request);
         }
 
         $method = strtolower($method);
@@ -51,22 +49,43 @@ trait ByMethodTrait
         }
     }
 
+
+    /**
+     * Handle HTTP OPTIONS request
+     */
+    public function options(
+        Request $request
+    ): Response {
+        return $this->handleUnknownMethod($request);
+    }
+
     /**
      * Handle unknown HTTP method
      */
     protected function handleUnknownMethod(
-        string $method
+        Request $request
     ): Response {
+        $method = $request->getMethod();
         $methods = [];
+        $route = $request->getAttribute('route');
 
         foreach (HarvestRequest::METHODS as $testMethod) {
-            if (method_exists($this, strtolower($testMethod))) {
-                $methods[] = $testMethod;
+            if (
+                !method_exists($this, strtolower($testMethod)) ||
+                (
+                    $route instanceof Route &&
+                    !$route->acceptsMethod($testMethod)
+                )
+            ) {
+                continue;
             }
+
+            $methods[] = $testMethod;
         }
 
         return Harvest::text('', $method === 'OPTIONS' ? 200 : 405, [
-            'allow' => implode(', ', $methods)
+            'Allow' => implode(', ', $methods),
+            'Access-Control-Allow-Methods' => implode(', ', $methods),
         ]);
     }
 }