+++ src/Harvest/Middleware/Greenleaf.php
@@ -9,10 +9,10 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Harvest\Middleware;
 
+use DecodeLabs\Greenleaf\Compiler\Hit;
 use DecodeLabs\Greenleaf\Context;
 use DecodeLabs\Greenleaf\Dispatcher;
 use DecodeLabs\Greenleaf\RouteNotFoundException;
-use DecodeLabs\Harvest;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Server\MiddlewareInterface as Middleware;
@@ -36,13 +36,7 @@ class Greenleaf implements
     public function handle(
         Request $request
     ): Response {
-        if ($request->getMethod() === 'OPTIONS') {
-            return Harvest::text('', 200, [
-                'allow' => 'OPTIONS, GET, HEAD, POST'
-            ]);
-        }
-
-        $hit = $this->context->matchIn($request, true);
+        $hit = $this->getHit($request);
 
         return $hit->getRoute()->handleIn(
             $this->context,
@@ -59,13 +53,7 @@ class Greenleaf implements
         Handler $next
     ): Response {
         try {
-            if ($request->getMethod() === 'OPTIONS') {
-                return Harvest::text('', 200, [
-                    'allow' => 'OPTIONS, GET, HEAD, POST'
-                ]);
-            }
-
-            $hit = $this->context->matchIn($request, true);
+            $hit = $this->getHit($request);
 
             return $hit->getRoute()->handleIn(
                 $this->context,
@@ -75,5 +63,16 @@ class Greenleaf implements
         } catch (RouteNotFoundException $e) {
             return $next->handle($request);
         }
+    }
+
+    /**
+     * Perform routing
+     */
+    protected function getHit(
+        Request &$request
+    ): Hit {
+        $hit = $this->context->matchIn($request, true);
+        $request = $request->withAttribute('route', $hit->getRoute());
+        return $hit;
     }
 }